Refactor database init helpers and restructure app entrypoint
