+++ app/db/__init__.py
@@ -1,0 +1,4 @@
+"""Database utilities."""
+from .init import bootstrap_schema, init_db
+
+__all__ = ["bootstrap_schema", "init_db"]
