+++ auth.py
@@ -1,17 +1,7 @@
 from __future__ import annotations
-from passlib.context import CryptContext
 from sqlalchemy.orm import Session
 from models import User
-
-pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
-
-
-def hash_password(password: str) -> str:
-    return pwd_context.hash(password)
-
-
-def verify_password(plain_password: str, password_hash: str) -> bool:
-    return pwd_context.verify(plain_password, password_hash)
+from app.core.security import hash_password, verify_password, pwd_context
 
 
 def get_user_by_username(db: Session, username: str) -> User | None:
